# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a doc for compass_mini_api.",
        "title": "compass_mini_api",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/feature/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Feature List Items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature"
                ],
                "summary": "Get Feature List Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Android",
                        "description": "entity",
                        "name": "entity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetFeatureListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/feature/sub/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Feature Sub List Items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feature"
                ],
                "summary": "Get Feature Sub List Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Android",
                        "description": "entity",
                        "name": "entity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id feature",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetFeatureSubResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/parameteritem/get_all_division": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Division",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParameterItem"
                ],
                "summary": "Get All Division",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\u003cbr\u003eexample string json: [{\"column\":\"name\",\"value\":\"Maintenance\",\"comparation\":\"=\"}]\u003cbr\u003eexample query encode: %5B%7B%22column%22%3A%22name%22%2C%22value%22%3A%22Maintenance%22%2C%22comparation%22%3A%22%3D%22%7D%5D",
                        "description": "filter by conditions with query encode value",
                        "name": "conditions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllDivisionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/change_password/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserChangePasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.GetAllDivisionResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GetAllDivisionResponse"
                    }
                }
            }
        },
...
        "dto.GetFeatureListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FeatureListResponse"
                    }
                }
            }
        },
        "dto.GetFeatureSubResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FeatureSubResponse"
                    }
                }
            }
        },
...
        "dto.UserChangePasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string",
                    "example": "Test12345@"
                },
                "old_password": {
                    "type": "string",
                    "example": "Test1234@"
                }
            }
        },
        "dto.UserChangePasswordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.FeatureListResponse": {
            "type": "object",
            "properties": {
                "allowadd": {
                    "type": "boolean"
                },
                "allowdelete": {
                    "type": "boolean"
                },
                "allowprint": {
                    "type": "boolean"
                },
                "allowupdate": {
                    "type": "boolean"
                },
                "allowview": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.FeatureSubResponse": {
            "type": "object",
            "properties": {
                "allowadd": {
                    "type": "boolean"
                },
                "allowdelete": {
                    "type": "boolean"
                },
                "allowprint": {
                    "type": "boolean"
                },
                "allowupdate": {
                    "type": "boolean"
                },
                "allowview": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.GetAllDivisionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isactive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parameter": {
                    "type": "string"
                },
                "parameterid": {
                    "type": "integer"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "error_id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": true
                }
            }
        },
        "response.errorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    }
}